Update and rename 2.8_withPCR.py to clip_template_PCR_2.8.py
Changed return tip to drop tip. Changed thermocycler module to take DEESTINATION_PLATE_TYPE, to make it more consistent with original template and to make it easier to change plate type. Changed to p10 pipette and a different tube rack. Removed unecessary comments, added more helpful comments.
